Move opamp client to a path that may resemble upstream
