+++ src/opentelemetry/_opamp/exceptions.py
