+++ src/opentelemetry/_opamp/transport/exceptions.py
